requirements.txt
asgiref==3.5.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.4
defusedxml==0.7.1
Django==3.2.21
django-allauth==0.54.0
django-cors-headers==3.14.0
django-crispy-forms==1.14.0
django-environ==0.10.0
django-filter==22.1
djangorestframework==3.14.0
ffmpeg-python==0.2.0
future==0.18.3
get-video-properties==0.1.1
idna==3.4
inotify==0.2.10
nose==1.3.7
oauthlib==3.2.2
Pillow==9.3.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.6
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
tinytag==1.9.0
urllib3==1.26.15
